Refactor store state management into separate appState module
